% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_statistics.R
\name{active_histogram}
\alias{active_histogram}
\title{Create histogram of active vs inactive tools}
\usage{
active_histogram(database)
}
\arguments{
\item{database}{A dataframe containing tool information with Still.active. column}
}
\value{
A ggplot2 object
}
\description{
Create histogram of active vs inactive tools
}
