% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_infographics.R
\name{active_tools}
\alias{active_tools}
\title{Count active tools over time}
\usage{
active_tools(
  database,
  ownership = FALSE,
  precision = TRUE,
  thisyear = as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{database}{A dataframe containing tool information}

\item{ownership}{Logical. If TRUE, displays ownership type as a stacked bar chart. Default is FALSE.}

\item{precision}{Logical. If FALSE, doesn't distinguish between private industry types. Default is TRUE.}

\item{thisyear}{Numeric. The current year for calculating active tools. Default is current year.}
}
\value{
A ggplot2 object
}
\description{
Count active tools over time
}
