% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcharts.R
\name{binary_X_categorical_Y_chart}
\alias{binary_X_categorical_Y_chart}
\title{Create a bar chart of a binary variable (Yes/No) against a categorical variable}
\usage{
binary_X_categorical_Y_chart(
  database,
  bin_x,
  cat_y,
  title = "",
  x_title = "",
  legend_title = "",
  caption = NA
)
}
\arguments{
\item{bin_x}{binary variable}

\item{cat_y}{categorical variable}

\item{title}{Defaults to ""}

\item{x_title}{title, for the x axis, defaults to ""}

\item{legend_title}{defaults to ""}

\item{caption}{optional caption (e.g., "*Excluded 1 tool, open-source status = Pending.")}
}
\description{
Create a bar chart of a binary variable (Yes/No) against a categorical variable
}
