% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_statistics.R
\name{calculate_yoy_growth}
\alias{calculate_yoy_growth}
\title{Calculate year-on-year growth in active tools}
\usage{
calculate_yoy_growth(
  database,
  thisyear = as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{database}{A dataframe containing tool information}

\item{thisyear}{Numeric. The current year for calculating active tools. Default is current year.}
}
\value{
A dataframe with year-on-year growth metrics
}
\description{
Calculate year-on-year growth in active tools
}
\examples{
yoy <- calculate_yoy_growth(database)

}
