% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{load_oecd_fonts}
\alias{load_oecd_fonts}
\title{Load Official OECD Fonts for Charts}
\usage{
load_oecd_fonts(verbose = interactive())
}
\arguments{
\item{verbose}{Logical. Whether to print informative messages. Defaults to
TRUE in interactive sessions, FALSE otherwise.}
}
\description{
Loads a modified version of Arial Narrow from P: drive that works with R
(and also Python).
}
