% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic.R
\name{mosaic_bin2}
\alias{mosaic_bin2}
\title{Mosaic Plot Using String Column Names}
\usage{
mosaic_bin2(df, x, y, title = NA, show_pct = FALSE)
}
\arguments{
\item{df}{A data frame containing the variables to be plotted.}

\item{x}{A string giving the name of the categorical variable for the x-axis.}

\item{y}{A string giving the name of the categorical variable for the fill variable.}

\item{title}{Optional plot title. If \code{NA} (default), a title is generated automatically.}

\item{show_pct}{Logical; if \code{TRUE}, displays percentage labels on the plot.}
}
\value{
A \code{ggplot} object displaying the mosaic plot.
}
\description{
This function creates a mosaic plot showing the relationship between two categorical
variables, where the variable names are supplied as strings. This version is useful
for programmatic workflows such as loops or \code{purrr::map()}.
}
\examples{
\dontrun{
library(ggmosaic)
library(dplyr)

data <- data.frame(
  ToolUsed = sample(c("Yes", "No"), 100, replace = TRUE),
  Category = sample(c("A", "B", "C"), 100, replace = TRUE)
)

mosaic_bin2(data, "ToolUsed", "Category", show_pct = TRUE)
}

}
