% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_statistics.R
\name{plot_yoy_growth}
\alias{plot_yoy_growth}
\title{Plot year-on-year growth in active tools}
\usage{
plot_yoy_growth(yoy_df, metric = "absolute")
}
\arguments{
\item{yoy_df}{A dataframe from calculate_yoy_growth function}

\item{metric}{Character. Type of metric to plot: "absolute", "percentage", or "both". Default is "absolute".}
}
\value{
A ggplot2 object
}
\description{
Plot year-on-year growth in active tools
}
\examples{
yoy <- calculate_yoy_growth(database)
plot_yoy_growth(yoy)

}
