% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_statistics.R
\name{release_histogram}
\alias{release_histogram}
\title{Create histogram of tool release years}
\usage{
release_histogram(database)
}
\arguments{
\item{database}{A dataframe containing tool information with Year.of.release column}
}
\value{
A ggplot2 object
}
\description{
Create histogram of tool release years
}
