% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcharts.R
\name{yes_no_histogram}
\alias{yes_no_histogram}
\title{Create a simple bar chart for binary variables}
\usage{
yes_no_histogram(database, bin_x, title = "")
}
\arguments{
\item{bin_x}{the binary variable to be shown, a column name in the database}

\item{title}{chart title (default is no title)}
}
\description{
Create a simple bar chart for binary variables
}
