% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcharts.R
\name{categorical_X_binary_Y_chart}
\alias{categorical_X_binary_Y_chart}
\title{Create a bar chart of a categorical variable against a binary variable (Yes/No)}
\usage{
categorical_X_binary_Y_chart(
  database,
  cat_x,
  bin_y,
  title = "",
  x_title = "",
  legend_title = "",
  caption = NA,
  position = "stack"
)
}
\arguments{
\item{database}{A data frame containing the variables to plot.}

\item{cat_x}{Categorical variable for the x-axis.}

\item{bin_y}{Binary variable for the fill/color (Yes/No).}

\item{title}{Plot title. Defaults to "".}

\item{x_title}{X-axis label. Defaults to "".}

\item{legend_title}{Legend title. Defaults to "".}

\item{caption}{Optional caption. If NA (default), shows number of tools.}

\item{position}{Bar position: "stack" (default), "dodge", or "fill".}
}
\value{
A ggplot2 object.
}
\description{
Create a bar chart of a categorical variable against a binary variable (Yes/No)
}
\examples{
\dontrun{
categorical_X_binary_Y_chart(database, Type.of.tool, Open.source)
categorical_X_binary_Y_chart(database, Type.of.tool, Open.source,
                              position = "dodge")
}

}
