% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_statistics.R
\name{latest_update_histogram}
\alias{latest_update_histogram}
\title{Create histogram of latest update years}
\usage{
latest_update_histogram(database)
}
\arguments{
\item{database}{A dataframe containing tool information with Latest.update column}
}
\value{
A ggplot2 object
}
\description{
Create histogram of latest update years
}
